w5/pg-mcp/pyproject.toml
[project]
name = "pg-mcp"
version = "0.1.0"
description = "PostgreSQL MCP Server with natural language query support"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.2",
    "asyncpg>=0.31.0",
    "sqlglot>=28.5.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "openai>=2.15.0",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
    "httpx[socks]>=0.28.1",
    "prometheus-client>=0.20.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
tracing = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "testcontainers[postgres]>=4.14.0",
    "ruff>=0.14.11",
    "mypy>=1.19.1",
]

[project.scripts]
pg-mcp = "pg_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pg_mcp"]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "testcontainers[postgres]>=4.14.0",
    "ruff>=0.14.11",
    "mypy>=1.19.1",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
# Filter deprecation warning from testcontainers library internals
# See: https://github.com/testcontainers/testcontainers-python/issues/303
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning:testcontainers",
]
